move assets
