+++ src/client/assets/styles/variables.ts
